feat(package.json): add jsqr library as a dependency
refactor(App.vue): reorganize imports and remove unused code
feat(App.vue): add enum AppMode for switching between Create and Capture modes
feat(App.vue): add ref for appMode and capturedData
feat(App.vue): add toggleAppMode function to switch between Create and Capture modes
feat(App.vue): add useCapturedDataInCreateMode function to set captured data and switch to Create mode
feat(App.vue): add computed property isModeToggleDisabled to disable mode toggle
feat(App.vue): add ref for QRCodeCapture component and capturedData
feat(App.vue): add enum ExportMode for single and batch export modes
feat(App.vue): add ref for exportMode, dataStringsFromCsv, filteredDataStringsFromCsv
feat(App.vue): add ref for inputFileForBatchEncoding, fileInput, isValidCsv, ignoreHeaderRow
feat(App.vue): add ref for isExportingBatchQRs, isBatchExportSuccess, currentExportedQr

feat(App.vue): add mode toggle button for switching between Create and Capture modes

refactor(App.vue): remove unused code related to QR code export and styling
feat(App.vue): add conditional rendering based on app mode for QR code creation and capture

style(App.vue): refactor styles for vertical-border and icon-button classes

feat(QRCodeCapture.vue): add QR code capture component with camera and image
upload functionality, including capturing, processing, and displaying QR code
data, copying data to clipboard, and creating QR code with captured data.

feat(QRCodeCreate.vue): add QRCodeCreate component with functionality to
generate QR codes, support batch export, error correction levels, and
styling options.
