+++ package.json
@@ -28,6 +28,7 @@
     "dom-to-image": "^2.6.0",
     "dom-to-svg": "^0.12.2",
     "file-saver": "^2.0.5",
+    "jsqr": "^1.4.0",
     "jszip": "^3.10.1",
     "lucide-vue-next": "^0.321.0",
     "qr-code-styling": "^1.9.1",
